Rearrange overrides into modules
Squashes warnings and cleans up tests.
